Move code: encoding.py > encoding/
