+++ kissim/encoding/__init__.py
@@ -1,0 +1,5 @@
+"""
+kissim.encoding
+"""
+
+from .api import Fingerprint, FingerprintGenerator
